+++ client/src/App.js
@@ -1,7 +1,8 @@
 import Home from "./components/Home";
-import FindCard from "./components/FindCard";
+import StoreCard from "./components/StoreCard";
 import About from "./components/About";
 import NavBar from "./components/NavBar";
+import Directory from "./components/Directory";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import "./App.css";
 function App() {
@@ -11,8 +12,9 @@ function App() {
       <BrowserRouter>
         <Routes>
           <Route index element={<Home />} />
+          <Route path="directory" element={<Directory />} />
           <Route path="about" element={<About />} />
-          <Route path="find" element={<FindCard />} />
+          <Route path="find" element={<StoreCard />} />
         </Routes>
       </BrowserRouter>
     </div>
