+++ client/src/components/Directory.js
@@ -1,0 +1,9 @@
+function Directory() {
+  return (
+    <div>
+      <h1> Directory</h1>
+    </div>
+  );
+}
+
+export default Directory;
