+++ client/src/components/NavBar.js
@@ -11,14 +11,15 @@ function NavBar() {
     <Navbar bg="light" expand="lg">
       <Container>
         <Navbar.Brand href="/">
-          <img style={myImageStyle} src={logo} alt="Money Pot Logo" />{" "}
-          Instruction Bot
+          <img style={myImageStyle} src={logo} alt="Money Pot Logo" /> MIPS
+          Translator
         </Navbar.Brand>
         <Navbar.Toggle aria-controls="basic-navbar-nav" />
         <Navbar.Collapse id="basic-navbar-nav">
           <Nav className="me-auto">
             <Nav.Link href="/">Home</Nav.Link>
             <Nav.Link href="about">About</Nav.Link>
+            <Nav.Link href="directory">Directory</Nav.Link>
           </Nav>
         </Navbar.Collapse>
       </Container>
